Refactor loader module: update package structure and rename files for clarity
