+++ enigma-engine/src/enigma/engine/factory/ReflectorFactory.java
@@ -1,6 +1,6 @@
 package enigma.engine.factory;
 
-import enigma.engine.model.ReflectorSpec;
+import enigma.shared.spec.ReflectorSpec;
 import enigma.machine.reflector.Reflector;
 
 /**
