+++ enigma-engine/src/enigma/engine/factory/RotorFactory.java
@@ -1,6 +1,6 @@
 package enigma.engine.factory;
 
-import enigma.engine.model.RotorSpec;
+import enigma.shared.spec.RotorSpec;
 import enigma.machine.rotor.Rotor;
 
 /**
